# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/trips/{tripId}/packing": {
      "get": {
        "operationId": "PackingController_findAll",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "description": "Trip ID (UUID)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by category",
            "schema": {
              "example": "Documents",
              "type": "string"
            }
          },
          {
            "name": "isPacked",
            "required": false,
            "in": "query",
            "description": "Filter by packed status (true/false)",
            "schema": {
              "example": "true",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PackingItemResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List packing items for a trip",
        "tags": [
          "Packing"
        ]
      },
      "post": {
        "operationId": "PackingController_create",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "description": "Trip ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePackingItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackingItemResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a packing item for a trip",
        "tags": [
          "Packing"
        ]
      }
    },
    "/trips/{tripId}/packing/progress": {
      "get": {
        "operationId": "PackingController_getProgress",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "description": "Trip ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackingProgressResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get packing progress summary for a trip",
        "tags": [
          "Packing"
        ]
      }
    },
    "/trips/{tripId}/packing/{id}": {
      "patch": {
        "operationId": "PackingController_update",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "description": "Trip ID (UUID)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Packing item ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePackingItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackingItemResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Packing item not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a packing item",
        "tags": [
          "Packing"
        ]
      },
      "delete": {
        "operationId": "PackingController_remove",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "description": "Trip ID (UUID)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Packing item ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Packing item not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a packing item",
        "tags": [
          "Packing"
        ]
      }
    },
    "/trips/{tripId}/packing/suggestions": {
      "get": {
        "operationId": "PackingController_getSuggestions",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "description": "Trip ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackingSuggestionsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Trip not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get AI-suggested packing items based on trip type, climate, and duration",
        "tags": [
          "Packing"
        ]
      }
    },
    "/trips/{tripId}/packing/suggestions/accept": {
      "post": {
        "operationId": "PackingController_acceptSuggestions",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "description": "Trip ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptPackingSuggestionsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PackingItemResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Trip not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Accept selected packing suggestions and add them to the trip packing list",
        "tags": [
          "Packing"
        ]
      }
    },
...
  },
  "info": {
    "title": "Travel Tracker API",
    "description": "API documentation for Travel Tracker backend services",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter access token"
      }
    },
    "schemas": {
...
      "DeleteResponseDto": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "description": "Indicates whether the resource was deleted",
            "example": true
          },
          "id": {
            "type": "string",
            "description": "Deleted resource identifier",
            "format": "uuid",
            "example": "1f8002e3-8f52-43f7-9c8d-e3a2603c1348"
          }
        },
        "required": [
          "deleted",
          "id"
        ]
      },
...
      "PackingItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Packing item ID",
            "format": "uuid",
            "example": "5fb3737f-100d-4df4-8498-e56f6690beaf"
          },
          "name": {
            "type": "string",
            "description": "Packing item name",
            "example": "Passport"
          },
          "isPacked": {
            "type": "boolean",
            "description": "Packed status",
            "example": true
          },
          "quantity": {
            "type": "number",
            "description": "Quantity required",
            "example": 1
          },
          "category": {
            "type": "string",
            "description": "Category",
            "example": "documents"
          },
          "tripId": {
            "type": "string",
            "description": "Trip ID",
            "format": "uuid",
            "example": "34ea6ea7-4a32-4832-8e0a-c10ba4f06673"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp",
            "example": "2026-03-22T09:30:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp",
            "example": "2026-03-22T09:30:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "isPacked",
          "quantity",
          "tripId",
          "createdAt",
          "updatedAt"
        ]
      },
      "PackingProgressResponseDto": {
        "type": "object",
        "properties": {
          "packed": {
            "type": "number",
            "description": "Number of packed items",
            "example": 8
          },
          "total": {
            "type": "number",
            "description": "Total number of items",
            "example": 20
          },
          "percentage": {
            "type": "number",
            "description": "Packed percentage (0-100)",
            "example": 40
          }
        },
        "required": [
          "packed",
          "total",
          "percentage"
        ]
      },
      "CreatePackingItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Packing item name",
            "example": "Passport"
          },
          "isPacked": {
            "type": "boolean",
            "description": "Whether item has been packed",
            "default": false
          },
          "quantity": {
            "type": "number",
            "description": "Number of units needed",
            "default": 1,
            "minimum": 1
          },
          "category": {
            "type": "string",
            "description": "Optional category label",
            "example": "documents"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdatePackingItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Packing item name",
            "example": "Passport"
          },
          "isPacked": {
            "type": "boolean",
            "description": "Whether item has been packed",
            "default": false
          },
          "quantity": {
            "type": "number",
            "description": "Number of units needed",
            "default": 1,
            "minimum": 1
          },
          "category": {
            "type": "string",
            "description": "Optional category label",
            "example": "documents"
          }
        }
      },
      "PackingSuggestionsResponseDto": {
        "type": "object",
        "properties": {
          "suggestions": {
            "type": "object",
            "description": "Suggested packing items grouped by category",
            "example": {
              "Clothes": [
                {
                  "key": "Clothes::T-shirts",
                  "name": "T-shirts",
                  "category": "Clothes",
                  "quantity": 5
                }
              ],
              "Toiletries": [],
              "Electronics": [],
              "Documents": [],
              "Misc": []
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of suggested items",
            "example": 12
          }
        },
        "required": [
          "suggestions",
          "total"
        ]
      },
      "AcceptPackingSuggestionsDto": {
        "type": "object",
        "properties": {
          "keys": {
            "description": "List of suggestion keys to accept. Pass all keys from suggestions response to accept all.",
            "example": [
              "Clothes::T-shirts",
              "Documents::Passport"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "keys"
        ]
      }
    }
  }
}